Refactor: brand, brandLike 분리
