+++ src/main/java/org/example/sivillage/brand/application/BrandService.java
@@ -9,15 +9,6 @@ import java.util.List;
 
 public interface BrandService {
 
-    /**
-     * BrandService interface
-     * 1. addBrand 브랜드 생성
-     * 2. getBrandIdList 브랜드 ID 목록 조회
-     * 3. getBrandName 브랜드 정보 조회
-     * 4. getBrandLike 브랜드 좋아요 여부 조회
-     * 5. toggleBrandLike 브랜드 좋아요 토글
-     */
-
 
     /**
      * 1. addBrand 브랜드 생성
@@ -40,19 +31,5 @@ public interface BrandService {
      */
     GetBrandNameResponseDto getBrandName(Long brandId);
 
-    /**
-     * 4. getBrandLike 브랜드 좋아요 여부 조회
-     * @param brandId 브랜드 ID
-     * @param memberUuid 회원 UUID
-     * return GetBrandLikeResponseDto
-     */
-    GetBrandLikeResponseDto getBrandLike(Long brandId, String memberUuid);
 
-    /**
-     * 5. toggleBrandLike 브랜드 좋아요 토글
-     * @param brandId 브랜드 ID
-     * @param memberUuid 회원 UUID
-     * return void
-     */
-    void toggleBrandLike(Long brandId, String memberUuid);
 }
