+++ src/main/java/org/example/sivillage/brand/presentation/BrandController.java
@@ -9,7 +9,6 @@ import org.example.sivillage.auth.domain.AuthUserDetails;
 import org.example.sivillage.brand.application.BrandService;
 import org.example.sivillage.brand.dto.in.AddBrandRequestDto;
 import org.example.sivillage.brand.vo.in.AddBrandRequestVo;
-import org.example.sivillage.brand.vo.out.GetBrandLikeResponseVo;
 import org.example.sivillage.brand.vo.out.GetBrandNameResponseVo;
 import org.example.sivillage.global.common.response.BaseResponse;
 import org.example.sivillage.global.common.response.dto.IdListResponseDto;
@@ -79,32 +78,5 @@ public class BrandController {
         return new BaseResponse<>(
                 brandService.getBrandName(brandId).toVo()
         );
-    }
-
-    /**
-     * 4. getBrandLike 브랜드 좋아요 여부 조회
-     * @param brandId 브랜드 ID
-     * @param authUserDetails 인증된 사용자 정보
-     * return GetBrandLikeResponseVo
-     */
-    @Operation(summary = "브랜드 좋아요 여부 조회")
-    @GetMapping("/{brandId}/like")
-    public BaseResponse<GetBrandLikeResponseVo> getBrandLike(@PathVariable Long brandId, @AuthenticationPrincipal AuthUserDetails authUserDetails) {
-        return new BaseResponse<>(
-                brandService.getBrandLike(brandId, authUserDetails.getMemberUuid()).toVo()
-        );
-    }
-
-    /**
-     * 5. toggleBrandLike 브랜드 좋아요 토글
-     * @param brandId 브랜드 ID
-     * @param authUserDetails 인증된 사용자 정보
-     * return void
-     */
-    @Operation(summary = "브랜드 좋아요 토글")
-    @PutMapping("/{brandId}/like")
-    public BaseResponse<Void> toggleBrandLike(@PathVariable Long brandId, @AuthenticationPrincipal AuthUserDetails authUserDetails) {
-        brandService.toggleBrandLike(brandId, authUserDetails.getMemberUuid());
-        return new BaseResponse<>();
     }
 }
